requirements.txt
marshmallow>=3
flask
flask-restful
flask-migrate
flask-sqlalchemy
flask-marshmallow
flask-jwt-extended
marshmallow-sqlalchemy
python-dotenv
passlib
apispec[yaml]
apispec-webframeworks
tox
gunicorn
dremio-client
python-dotenv
sqlalchemy_dremio
psycopg2
numpy
sqlalchemy_jsonfield
boto3
pyodbc==4.0.32
